rgmanager: ra2rng.xsl: common params: prevent misleading combination
...both @__max_restarts and @__restart_expire_time have "nonzero" values,
otherwise, otherwise if either of the parameter is present, it has to have
"zero" value.

Every other occurrence of these two interconnected parameters for
particular resource is a configuration defect, as it could lead
to false assumption that one will take effect even without the
other defined and nonzero.

Reference:
>		if (restart_expire_time == 0 || max_restarts == 0)
>			return;
>		goto out_assign;
[https://git.fedorahosted.org/cgit/cluster.git/tree/rgmanager/src/daemons/restree.c?h=RHEL64#n538]

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>