rgmanager: ra2rng.xsl et al.: obsolete 1 pass in ccs_update_schema
... in a backward-compatible way, which is done by wrapping the output
of this as-of-now obsoleted pass by element from explicit non-RelaxNG
namespace, which is then, in the validation phase, completely skipped.

Separate patch for cluster.git to actually remove the pass in question
will be provided, but thanks to this backward compatibility, it is not
strictly required -- only informative notes about possibility of outdated
cman package will possibly occur.  The corresponding pass is also removed
directly in the local Makefile mirroring the part of ccs_update_schema
functionality so the resulting resources.rng does not contain the
now-redundant references.

[ It's nice to see how smart the design of Relax NG, allowing
  for such define-ala-mixin, is. ]

NB: the change "optional" -> "zeroOrMore" is because previously
    "optional" was previously pointless (as is chaining '?' + '*'
    in regexp), but now the nested "zeroOrMore" is promoted from
    the definition of CHILDREN block up to its user (the other
    users were already using "zeroOrMore" also before descending
    to such symbol, which was another pointless thing, FWIW)

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>