rgmanager: ra2rng.xsl et al.: enforce "forbid child" from RA metadata
As with previous change, the reflection of this change is yet to come
to cluster.git (ccs_update_schema script) via a separate patch.

There is a new file included in "DIST", ra2rng-postprocess.xsl, that
is to be appended  to a pipe with original ra2rng.xsl (both with xsltproc
as an effective actor, indeed), that in turn now requires some extra
decoration as can be found also in the respective Makefile.am.

In netfs.sh, the order of forbidden children in metadata now follows
alphabetical ordering, which is in-line with other such occurrences
in RAs metadata (and in turn, ip.sh and netfs.sh share the same
"forbidstring" rather than each its own despite being semantically
the same).

Also note that this provision to schema generation will not work
properly until libxslt bugfix [1] is propagated to the target
deployment.

[1] https://mail.gnome.org/archives/xslt/2013-December/msg00001.html

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>